requirements.txt
# Google Sheets API
gspread==6.1.4
google-auth==2.38.0

# Dashboard UI
streamlit==1.41.1

# Data processing
pandas==2.2.3

# Environment variables
python-dotenv==1.0.1
